Public int added

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
 
-    public double calculate(double massInKg, double heightInM) {
+    public int calculate(double massInKg, double heightInM) {
 
         int bmiIndex = (int) (massInKg / (heightInM * heightInM));
